requirements.txt
streamlit>=1.24.0
openai
python-dotenv
